feat(invalid-tiles): Preserve and visualize invalid OTBM content
Added preservation and visualization of invalid/unresolved OTBM tile data instead of discarding it. Introduced InvalidOTBMItemData and InvalidZoneState structs to capture raw bytes and node structures for missing definitions. New menu actions 'Cleanup invalid tiles' and 'Cleanup invalid zones' allow targeted removal. Toggle visibility via 'Show Invalid Tiles' (red/orange overlays) and 'Show Invalid Zones' (magenta overlays). Expanded mapflags from uint16_t to uint32_t to support unknown flag bits. Serialization now writes preserved raw data back to OTBM on save.

+++ source/lua/lua_api_tile.cpp
@@ -393,7 +393,7 @@ namespace LuaAPI {
 			"isPvpZone", sol::property([](Tile* tile) { return tile && (tile->getMapFlags() & TILESTATE_PVPZONE); }),
 
 			// Map flags
-			"mapFlags", sol::property([](Tile* tile) -> uint16_t { return tile ? tile->getMapFlags() : 0; }, [](Tile* tile, uint16_t flags) {
+			"mapFlags", sol::property([](Tile* tile) -> uint32_t { return tile ? tile->getMapFlags() : 0; }, [](Tile* tile, uint32_t flags) {
 					if (tile) {
 						markTileForUndo(tile);
 						tile->setMapFlags(flags);
